#224: extract shared OnCloseFileSource wrapper
Unify StagedFileSource (iosMain) and FolderChildFileSource (appleMain)
into one thread-safe idempotent-close delegate in appleMain; both call
sites reuse it. Behavior preserved; close() guard upgraded to atomic.
